Restructures main.py to organize folders and filenames
We will need to do more work to clean up this file but this is all I thought made sense right now. I ran into errors where the directories didn't exist on my computer so I added a check to see if they exist and create them if not. Filename errors are generated separately within the code when they get executed right now but we should probably move up the file exists check to this part of the code. We may also want to consider separating these filenames out into a config file. Will add this info in a new issue as well.